Use config port for mock api calls in functional tests
The port is configurable in some places, but the port used for address lookup api calls isn't using the same config, so it can't actually be configured without failing.

Instead use the configured port everywhere so that cahnges to config are respected and the tests run correctly irrespective of the port configuration.